p4i integration Integrate 230897 from 4.2 to main: Fixes:    Ambiguous call to QAbstractItemModel::createIndex().
[git-p4: depot-paths = "//depot/qt/main/": change = 230898]

+++ tests/auto/qtreeview/tst_qtreeview.cpp
@@ -111,7 +111,7 @@ public:
         if (row < 0 || column < 0 || column >= cols || row >= rows) {
             return QModelIndex();
         }
-        qint64 level = parent.isValid() ? parent.internalId() : -1;
+        qint32 level = parent.isValid() ? parent.internalId() : -1;
         QModelIndex i = createIndex(row, column, ++level);
 	parentHash[i] = parent;
 	return i;
